requirements.txt
Flask==2.0.1
imbalanced_learn==0.8.1
joblib==1.2.0
numpy==1.21.2
pandas==1.3.3
python-dotenv==0.19.0
scikit-learn==0.24.2
Werkzeug==2.0.1
xgboost==2.1.4
reportlab==3.6.8